+++ pbp/hmb_gen/process_helper.py
@@ -164,8 +164,6 @@ class ProcessHelper:
         if attrs_uri:
             self.log.info(f"Loading {what} attributes from {attrs_uri=}")
             filename = self.file_helper.get_local_filename(attrs_uri)
-            if os.name == "nt" and filename is not None:
-                filename = filename[3:]
             if filename is not None:
                 with open(filename, "r", encoding="UTF-8") as f:
                     res = parse_attributes(f.read(), pathlib.Path(filename).suffix)