+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 2025-11
 
 - 1.8.7:
-
+   - Fixed issue where relative paths were not handled properly on Windows.
    -  hmb-gen: Loading/completion of global attributes are now performed
       with before the actual processing.